+++ src/libYARP_OS/src/BufferedConnectionWriter.cpp
@@ -27,7 +27,7 @@ bool BufferedConnectionWriter::applyConvertTextMode() {
             sos.write(m.usedBytes());
         }
         const std::string& str = sos.str();
-        b.fromBinary(str.c_str(), str.length());
+        b.fromBinary(str.c_str(), (int)str.length());
         std::string replacement = b.toString() + "\n";
         for (size_t i=0; i<lst.size(); i++) {
             delete lst[i];
